+++ src/index.ts
@@ -12,11 +12,11 @@ import type {
   Context,
   ExecutionMethod,
   Finished,
-  RecursivePartial,
   Result,
   Error as ResultError,
   SVMProgram,
 } from './types';
+import type { RecursivePartial } from './utils/typeUtils';
 import type { ModuleContext, ImportOptions } from './modules/moduleTypes';
 import { assemble } from './vm/svml-assembler';
 import { compileToIns } from './vm/svml-compiler';
