+++ src/types.ts
@@ -398,26 +398,6 @@ export type TypeEnvironment = {
 }[];
 
 /**
- * Helper type to recursively make properties that are also objects
- * partial
- *
- * By default, `Partial<Array<T>>` is equivalent to `Array<T | undefined>`. For this type, `Array<T>` will be
- * transformed to Array<Partial<T>> instead
- */
-export type RecursivePartial<T> =
-  T extends Array<any>
-    ? Array<RecursivePartial<T[number]>>
-    : // Apparently typescript considers functions assignable
-      // to records?
-      T extends (...args: any[]) => any
-      ? T
-      : T extends Record<any, any>
-        ? Partial<{
-            [K in keyof T]: RecursivePartial<T[K]>;
-          }>
-        : T;
-
-/**
  * Utility type for selecting extracting the specific Node type when provided
  * the type string.
  */
